unify test configuration method
Remove the individual test overrides for each of the loopcount,
usercount, delay, session length options.

Most of these tests are meant to be run independently so there
is no reason that they each need their own identical value.
If there is a need to layer load from multiple test suites then
replicas of each 'type' should be created each generating some amount
of user activity and load.
This should allow for much greater flexibility when it comes to
testing the Zimbra server.